Modified GUI so GUI elements always destroy their children when they are destroyed Added methods and UI for modifying arrays in Inspector (WIP) Added updatePlugin functions to plugins Mono method calls will always forward exceptions to C++ now Fixed a Win32 compilation error due to incorrect function call type

+++ BansheeCore/Source/BsPixelUtil.cpp
@@ -843,7 +843,7 @@ namespace BansheeEngine
 
 		virtual void beginImage(int size, int width, int height, int depth, int face, int miplevel)
 		{ 
-			assert(miplevel >= 0 && miplevel < buffers.size());
+			assert(miplevel >= 0 && miplevel < (int)buffers.size());
 			assert(size == buffers[miplevel]->getConsecutiveSize());
 
 			activeBuffer = buffers[miplevel];
